udd: do not use PATH when execcing dd with privileges
Instead of searching for a program called 'dd' on the PATH, we try two
hardcoded directories. POSIX seems to ensure that the system dd program
is located in /usr/bin or in /bin, so we can stop relying on the PATH
list to not be rogue.

This makes a difference if a user process gains privileges and passes a
rogue PATH so we execute a malicious dd program.